move code to a top-level package (factom_did)
