+++ factom_did/client/keys/management.py
@@ -1,19 +1,20 @@
 import base58
 
-from client.constants import ENTRY_SCHEMA_V100
-from client.enums import DIDKeyPurpose, KeyType
-from client.keys.abstract import AbstractDIDKey
+from factom_did.client.constants import ENTRY_SCHEMA_V100
+from factom_did.client.enums import KeyType
+from factom_did.client.keys.abstract import AbstractDIDKey
 
 
-class DIDKey(AbstractDIDKey):
+class ManagementKey(AbstractDIDKey):
     """
-    Application-level key, which can be used for authentication, signing requests, encryption, decryption, etc.
+    A key used to sign updates for an existing DID.
 
     Attributes
     ----------
     alias: str
-    purpose: DIDKeyPurpose or DIDKeyPurpose[]
-        Shows what purpose(s) the key serves. (PublicKey, AuthenticationKey or both)
+    priority: int
+        A non-negative integer showing the hierarchical level of the key. Keys with lower priority override keys with
+        higher priority.
     key_type: KeyType
     controller: str
     priority_requirement: int, optional
@@ -24,7 +25,7 @@ class DIDKey(AbstractDIDKey):
     def __init__(
         self,
         alias,
-        purpose,
+        priority,
         key_type,
         controller,
         priority_requirement=None,
@@ -35,32 +36,21 @@ class DIDKey(AbstractDIDKey):
             alias, key_type, controller, priority_requirement, public_key, private_key
         )
 
-        if type(purpose) is list:
-            purpose_l = purpose
-        else:
-            purpose_l = [purpose]
+        if priority < 0:
+            raise ValueError("Priority must be a non-negative integer.")
 
-        for purpose_type in purpose_l:
-            if purpose_type not in {
-                DIDKeyPurpose.PublicKey,
-                DIDKeyPurpose.AuthenticationKey,
-            }:
-                raise ValueError(
-                    "Purpose must contain only valid DIDKeyPurpose values."
-                )
-
-        self.purpose = purpose_l
+        self.priority = priority
 
     def __eq__(self, other):
         if self.__class__ is other.__class__:
-            return super().__eq__(other) and self.purpose == other.purpose
+            return super().__eq__(other) and self.priority == other.priority
         return NotImplemented
 
     def __hash__(self):
         return hash(
             (
                 self.alias,
-                "".join(map(lambda x: x.value, self.purpose)),
+                self.priority,
                 self.key_type,
                 self.controller,
                 self.priority_requirement,
@@ -71,12 +61,12 @@ class DIDKey(AbstractDIDKey):
 
     def __repr__(self):
         return (
-            "<{}.{}(alias={}, purpose={}, key_type={}, controller={}, "
+            "<{}.{}(alias={}, priority={}, key_type={}, controller={}, "
             "priority_requirement={})>".format(
                 self.__module__,
                 type(self).__name__,
                 self.alias,
-                self.purpose,
+                self.priority,
                 self.underlying,
                 self.controller,
                 self.priority_requirement,
@@ -86,7 +76,7 @@ class DIDKey(AbstractDIDKey):
     def to_entry_dict(self, did, version=ENTRY_SCHEMA_V100):
         if version == ENTRY_SCHEMA_V100:
             d = super().to_entry_dict(did)
-            d["purpose"] = list(map(lambda x: x.value, self.purpose))
+            d["priority"] = self.priority
             return d
         else:
             raise NotImplementedError("Unknown schema version: {}".format(version))
@@ -94,9 +84,9 @@ class DIDKey(AbstractDIDKey):
     @staticmethod
     def from_entry_dict(entry_dict, version=ENTRY_SCHEMA_V100):
         if version == ENTRY_SCHEMA_V100:
-            return DIDKey(
+            return ManagementKey(
                 alias=entry_dict["id"].split("#")[-1],
-                purpose=list(map(DIDKeyPurpose.from_str, entry_dict.get("purpose"))),
+                priority=entry_dict["priority"],
                 key_type=KeyType.from_str(entry_dict["type"]),
                 controller=entry_dict["controller"],
                 priority_requirement=entry_dict.get("priorityRequirement"),
