+++ factom_did/resolver/exceptions.py
